requirements.txt
numpy
Pillow
scikit-learn
albumentations